New Zealand Spatial Features

The nzsf package provides two main categories of important functions: get and plot. Several other helper functions are also available, such as the proj functions which can be used to easily specify a useful projection to use.

Get functions

The get functions obtain shapefiles using a call to data and do some processing of that data (e.g. transform to a different projection). These functions return a spatial features object.

Plot functions

The plot functions include a get function and an appropriate ggplot2 geom to plot the shapefile.

See also

Author

Maintainer: Darcy Webber darcy.webber@quantifish.co.nz